fix: merge conflicts and various minor refactoring
