+++ assets/wizards/newspack/views/dashboard/index.tsx
@@ -13,11 +13,11 @@ import { render } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { GlobalNotices, Footer, Notice, Wizard } from '../../components/src';
+import { GlobalNotices, Footer, Notice, Wizard } from '../../../../components/src';
+import sections from './sections';
+import BrandHeader from '../../components/brand-header';
+import QuickActions from '../../components/quick-actions';
 import './style.scss';
-import sections from './components/sections';
-import BrandHeader from './components/brand-header';
-import QuickActions from './components/quick-actions';
 
 const {
 	newspack_aux_data: { is_debug_mode: isDebugMode = false },
