+++ assets/wizards/newspack/views/dashboard/style.scss
@@ -3,7 +3,7 @@
  */
 
 @use '~@wordpress/base-styles/colors' as wp-colors;
-@use '../../shared/scss/colors';
+@use '../../../../shared/scss/colors';
 
 :root {
 	--np-font-size: 16px;
